split juster.ts to separate files
