+++ packages/juster-sdk/src/index.ts
@@ -1,3 +1,4 @@
-export * from "./juster";
+export * from "./juster/core";
+export * from "./juster/pool";
 export * from "./estimates";
 export * from "./types";
